applied review comments
